Fix compile warning in Catalog
git-svn-id: http://192.168.20.11/sequoiadb/trunk@25097 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cat/catCommon.cpp
@@ -3725,7 +3725,7 @@ namespace engine
             beginBound = endBound ;
             endBound = endBound + avgBound ;
             ++ itemID ;
-            if ( grpSize - 1 == itemID )
+            if ( grpSize - 1 == (UINT32)itemID )
             {
                endBound = totalBound ;
             }
